change StepMeter to WizardHeader and pull meter into own component
