+++ app/src/atoms/StepMeter/StepMeter.stories.tsx
@@ -1,13 +1,5 @@
 import * as React from 'react'
-import {
-  Flex,
-  DIRECTION_ROW,
-  JUSTIFY_FLEX_END,
-  DIRECTION_COLUMN,
-  COLORS,
-  SPACING,
-} from '@opentrons/components'
-import { PrimaryButton, SecondaryButton } from '../buttons/index'
+
 import { StepMeter } from './index'
 
 import type { Story, Meta } from '@storybook/react'
@@ -25,39 +17,4 @@ export const Primary = Template.bind({})
 Primary.args = {
   totalSteps: 5,
   currentStep: 2,
-  title: 'Tip Length Calibration',
-  body: (
-    <Flex flexDirection={DIRECTION_COLUMN}>
-      {'this is the body part of the StepMeter'}
-    </Flex>
-  ),
-  footer: (
-    <Flex flexDirection={DIRECTION_ROW} justifyContent={JUSTIFY_FLEX_END}>
-      <Flex marginRight={SPACING.spacing4}>
-        <PrimaryButton backgroundColor={COLORS.blue}>{'Back'}</PrimaryButton>
-      </Flex>
-      <SecondaryButton color={COLORS.blue}>{'Proceed'}</SecondaryButton>
-    </Flex>
-  ),
-  showStepCount: true,
-}
-
-export const NoStepCount = Template.bind({})
-NoStepCount.args = {
-  totalSteps: 5,
-  currentStep: 0,
-  title: 'Tip Length Calibration',
-  body: (
-    <Flex flexDirection={DIRECTION_COLUMN}>
-      {'this is the body part of the StepMeter'}
-    </Flex>
-  ),
-  footer: (
-    <Flex flexDirection={DIRECTION_ROW} justifyContent={JUSTIFY_FLEX_END}>
-      <Flex marginRight={SPACING.spacing4}>
-        <PrimaryButton backgroundColor={COLORS.blue}>{'Back'}</PrimaryButton>
-      </Flex>
-      <SecondaryButton color={COLORS.blue}>{'Proceed'}</SecondaryButton>
-    </Flex>
-  ),
 }
